Add keywords.go and move some of the funcs and types
